third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-stay-in-overflow-div-expected.txt

a
b
c
PASS successfullyParsed is true

TEST COMPLETE
PASS gFocusedDocument.activeElement.getAttribute("id") is "a"
PASS gFocusedDocument.activeElement.getAttribute("id") is "b"
PASS gFocusedDocument.activeElement.getAttribute("id") is "c"
PASS gFocusedDocument.activeElement.getAttribute("id") is "b"
PASS gFocusedDocument.activeElement.getAttribute("id") is "b"
PASS gFocusedDocument.activeElement.getAttribute("id") is "a"
